move octree to lib, remove unness atomicOperations
