+++ src/ts/app/view/toolbar.ts
@@ -2,7 +2,6 @@
 
 import * as _ from 'underscore';
 import * as Backbone from 'backbone';
-import { atomic } from '../model/atomic';
 
 export const LandmarkSizeSlider = Backbone.View.extend({
 
@@ -24,11 +23,11 @@ export const LandmarkSizeSlider = Backbone.View.extend({
         return this;
     },
 
-    changeLandmarkSize: atomic.atomicOperation(function (event) {
+    changeLandmarkSize: function (event) {
         this.model.set(
             "landmarkSize",
             Math.max(Number(event.target.value) / 100, 0.05));
-    })
+    }
 });
 
 export const TextureToggle = Backbone.View.extend({
